# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "E3Di & Jaan LED Screen IoT Ad-Network Platform API",
    "description": "Comprehensive REST & IoT API documentation for the E3Di Smart Pole, Digital Billboard Network, Edge Gateways, CCTV surveillance, and Emergency SOS infrastructure.",
    "version": "2.0.0",
    "contact": {
      "name": "E3Di Platform Support",
      "url": "https://e3di.org"
    }
  },
  "servers": [
    {
      "url": "https://api.e3di.org",
      "description": "Production LED CPU API Server"
    },
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server (Port 5000)"
    },
    {
      "url": "http://localhost:5001",
      "description": "Alternative Local Server (Port 5001)"
    }
  ],
...
  "paths": {
...
    "/api/admin/campaigns/counts": {
      "get": {
        "tags": ["Admin Campaign Management"],
        "summary": "Get Campaign Status Tallies",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Counts for pending, approved, active, rejected, and completed" }
        }
      }
    },
    "/api/admin/campaigns/pending": {
      "get": {
        "tags": ["Admin Campaign Management"],
        "summary": "List Pending Review Campaigns",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Pending campaigns" } }
      }
    },
    "/api/admin/campaigns/active": {
      "get": {
        "tags": ["Admin Campaign Management"],
        "summary": "List Currently Airing Campaigns",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Active campaigns" } }
      }
    },
    "/api/admin/campaigns/{id}/approve": {
      "patch": {
        "tags": ["Admin Campaign Management"],
        "summary": "Approve Campaign for Airing",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Campaign approved and pushed to screens" } }
      }
    },
    "/api/admin/campaigns/{id}/reject": {
      "patch": {
        "tags": ["Admin Campaign Management"],
        "summary": "Reject Campaign",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object", "properties": { "rejectionReason": { "type": "string" } } } } }
        },
        "responses": { "200": { "description": "Campaign rejected" } }
      }
    },
    "/api/admin/campaigns/{id}/revoke": {
      "patch": {
        "tags": ["Admin Campaign Management"],
        "summary": "Revoke Approved Campaign",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Campaign revoked and stopped immediately" } }
      }
    },
    "/api/admin/campaigns/corridor-state": {
      "get": {
        "tags": ["Admin Campaign Management"],
        "summary": "Real-time Corridor Broadcast Playback State",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Broadcast synchronization state across corridor poles" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
...
  }
}